Refactoring some tests
